+++ resources/merchant.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -11,22 +11,26 @@ blp = Blueprint("Merchants", __name__, description="Operations on merchants")
 
 @blp.route("/merchant/<string:merchant_id>")
 class Merchant(MethodView):
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.response(200, MerchantSchema)
     def get(self, customer_id):
         merchant = MerchantModel.query.get_or_404(merchant_id)
 
         return merchant
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def delete(self, merchant_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+        
         merchant = MerchantModel.query.get_or_404(merchant_id)
         db.session.delete(merchant)
         db.session.commit()
 
         return {"Message": "Merchant deleted."}
 
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.arguments(MerchantUpdateSchema)
     @blp.response(200, MerchantSchema)
     def put(self, merchant_data, merchant_id):
@@ -44,15 +48,19 @@ class Merchant(MethodView):
 
 @blp.route("/merchant")
 class MerchantList(MethodView):
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.response(200, MerchantSchema(many=True))
     def get(self):
         return MerchantModel.query.all()
 
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.arguments(MerchantSchema)
     @blp.response(201, MerchantSchema)
     def post(self, merchant_data):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+        
         merchant = MerchantModel(**merchant_data)
 
         try:
