+++ resources/score.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from datetime import datetime
@@ -13,22 +13,26 @@ blp = Blueprint("Score", __name__, description="Operations on score")
 
 @blp.route("/score/<string:score_id>")
 class Score(MethodView):
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.response(200, ScoreSchema)
     def get(self, score_id):
         score = ScoreModel.query.get_or_404(score_id)
 
         return score
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def delete(self, score_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         score = ScoreModel.query.get_or_404(score_id)
         db.session.delete(score)
         db.session.commit()
 
         return {"Message": "Score deleted."}
 
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.arguments(ScoreUpdateSchema)
     @blp.response(200, ScoreSchema)
     def put(self, score_data, score_id):
@@ -48,28 +52,32 @@ class Score(MethodView):
 
 @blp.route("/score")
 class ScoreList(MethodView):
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.response(200, ScoreSchema(many=True))
     def get(self):
         return ScoreModel.query.all()
 
 @blp.route("/score/t/<string:transaction_id>")
 class ScoreTransaction(MethodView):
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.response(200, ScoreSchema(many=True))
     def get(self, transaction_id):
         scores = ScoreModel.query.filter_by(transaction_id=transaction_id).all_or_404()
         return scores
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def delete(self, transaction_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         scores = ScoreModel.query.filter_by(transaction_id=transaction_id).all_or_404()
         for score in scores:
             db.session.delete(score)
         db.session.commit()
         return {"Message": "Scores deleted."}
     
-    @jwt_required()
+    @jwt_required(refresh=True)
     @blp.arguments(ScoreSchema)
     @blp.response(201, ScoreSchema)
     def post(self, score_data):
